// src/app/shared/model/Ireels.ts
export interface Ireels {
    
        title: string;
        likeCount: number;
        isLike: boolean;
        id: number;
  
}
export interface IindianYouTubers {
        name: string;
        subscribers: number;
        id: number;
        isLikeCount :number;
        isDislikeCount :number;
        isLike :boolean;

    }